+++ include/linAlg.hpp
@@ -519,6 +519,23 @@ namespace MathLib {
         }
 
         /**
+         * @brief Calculate the inner product of given 2 vectors "x" and "y".
+         *
+         * @tparam T the number type of the elements of "x" and "y"
+         * @param[in] N the length of vector "x" and "y"
+         * @param[in] x the vector "x"
+         * @param[in] y the vector "y"
+         * @return the inner product of "x" and "y"
+         */
+        template <typename T>
+        T innerProd(const size_t N, const T *x, const T*y) {
+            constexpr T ZERO = static_cast<T>(0);
+            T sum = ZERO;
+            for (size_t n=0; n<N; ++n) {sum += x[n]*y[n];}
+            return sum;
+        }
+
+        /**
          * @brief Calculates Hermitian inner product of given 2 complex vectors "x", "y".
          * Hermitian inner product of "x" and "y" is defined as "<x^*, y>"", where "^*" represents conjugate and "<,>" represents inner product.
          *